3.接口用例脚本执行框架/api_autotest/requirements.txt
pytest>=7.0.0
pytest-html>=3.1.1
allure-pytest>=2.9.45
requests>=2.26.0
pyyaml>=6.0
pandas>=1.3.0
python-dotenv>=0.19.0
openpyxl>=3.0.9
